+++ CHANGELOG.md
@@ -13,6 +13,7 @@ All notable changes to this project will be documented in this file.
 ### Quality
 
 - Configure Gradle JVM Test Suite Plugin. (#448)
+- Remove `ReplicateActionResponse` and `ReplicateTaskSummary` classes. (#452)
 
 ### Dependency Upgrades