+++ CHANGELOG.md
@@ -26,6 +26,7 @@ All notable changes to this project will be documented in this file.
 
 - Upgrade to Gradle 8.7. (#449)
 - Upgrade to Spring Boot 2.7.18. (#450)
+- Upgrade to `iexec-commons-poco` 4.1.0. (#457)
 
 ## [[8.4.0]](https://github.com/iExecBlockchainComputing/iexec-common/releases/tag/v8.4.0) 2024-02-27